# runtime_stack/condenser/docs/swagger.yaml
definitions:
...
  utils.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/bottle/{bottleId}:
    get:
      description: get bottle detail by id or name
      parameters:
      - description: Bottle ID or Name
        in: path
        name: bottleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get bottle detail
      tags:
      - bottles
  /v1/bottle/{bottleId}/actions/{action}:
    post:
      description: start/stop/remove bottle
      parameters:
      - description: Bottle ID or Name
        in: path
        name: bottleId
        required: true
        type: string
      - description: Action
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: perform bottle action
      tags:
      - bottles
...
swagger: "2.0"